requirements.txt
numpy>=1.20.0
pandas>=1.3.0
scipy>=1.7.0
pyyaml>=5.4.0
pytest>=6.0.0
hypothesis>=6.0.0